requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
pytest>=7.4.0

